+++ src/collective/documentgenerator/content/pod_template.py
@@ -7,6 +7,11 @@ from collective.documentgenerator.utils import compute_md5
 
 from collective.z3cform.datagridfield import DataGridFieldFactory
 from collective.z3cform.datagridfield import DictRow
+from collective.documentgenerator.config import NEUTRAL_FORMATS
+from collective.documentgenerator.config import ODS_FORMATS
+from collective.documentgenerator.config import ODT_FORMATS
+from collective.documentgenerator.config import POD_FORMATS
+from collective.documentgenerator.content.style_template import StyleTemplate
 
 from plone import api
 from plone.autoform import directives as form
@@ -15,16 +20,22 @@ from plone.formwidget.namedfile import NamedFileWidget
 from plone.namedfile.field import NamedBlobFile
 from plone.supermodel import model
 
+from z3c.form import validator
 from z3c.form.browser.select import SelectWidget
 
 from zope import schema
+from zope.component import provideAdapter
 from zope.component import queryAdapter
 from zope.component import queryMultiAdapter
 from zope.interface import implements
-from zope.schema._messageid import _ as zope_message_factory
+from zope.interface import Interface
+from zope.interface import Invalid
 
 import logging
-import zope
+
+from Products.CMFPlone.utils import safe_unicode
+
+
 logger = logging.getLogger('collective.documentgenerator: PODTemplate')
 
 
@@ -55,6 +66,7 @@ class PODTemplate(Item):
     """
 
     implements(IPODTemplate)
+    meta_type = 'PODTemplate'
 
     def get_file(self):
         return self.odt_file
@@ -103,7 +115,7 @@ class PODTemplate(Item):
         return has_been_modified
 
 
-class IMergeTemplatesRowSchema(zope.interface.Interface):
+class IMergeTemplatesRowSchema(Interface):
     """
     Schema for DataGridField widget's row of field 'merge_templates'
     """
@@ -119,19 +131,42 @@ class IMergeTemplatesRowSchema(zope.interface.Interface):
     )
 
 
+class PodFormatsValidator(validator.SimpleFieldValidator):
+    """ z3c.form validator class for international phone numbers """
+
+    def validate(self, selected_formats):
+        # thanks to widget, we get just-loaded file.filename.
+        widgets = self.view.widgets
+        current_filename = u""
+        for element in widgets.items():
+            if element[0] == 'odt_file':
+                current_filename = safe_unicode(element[1].filename)
+        if current_filename:
+            FORMATS_DICT = {'ods': ODS_FORMATS + NEUTRAL_FORMATS,
+                            'odt': ODT_FORMATS + NEUTRAL_FORMATS}
+            extension = current_filename.split('.')[-1]
+            authorise_element_list = FORMATS_DICT[extension]
+            authorise_extension_list = [elem[0] for elem in authorise_element_list]
+            if not selected_formats:
+                raise Invalid(_(u"No format selected"))
+            for element in selected_formats:
+                if element not in authorise_extension_list:
+                    elem = self.get_invalid_error(element)
+                    error_message = _(u"element_not_valid",
+                                      default=u"Element ${elem} is not valid for .${extension} template : \"${template}\"",
+                                      mapping={u"elem": elem, u"extension": extension, u"template": current_filename})
+                    raise Invalid(error_message)
+
+    def get_invalid_error(self, extension):
+        for element in POD_FORMATS:
+            if element[0] == extension:
+                return element[1]
+
+
 class IConfigurablePODTemplate(IPODTemplate):
     """
     ConfigurablePODTemplate dexterity schema.
     """
-
-    def pod_formats_constraint(value):
-        """
-        By default, it seems that 'required' is not correctly validated
-        so we double check that the field is not empty...
-        """
-        if not value:
-            raise zope.interface.Invalid(zope_message_factory(u'Required input is missing.'))
-        return True
 
     pod_formats = schema.List(
         title=_(u'Available formats'),
@@ -139,7 +174,6 @@ class IConfigurablePODTemplate(IPODTemplate):
         value_type=schema.Choice(source='collective.documentgenerator.Formats'),
         required=True,
         default=['odt', ],
-        constraint=pod_formats_constraint,
     )
 
     form.widget('pod_portal_types', SelectWidget, multiple='multiple', size=15)
@@ -169,6 +203,12 @@ class IConfigurablePODTemplate(IPODTemplate):
         default=[],
     )
 
+# Set conditions for which fields the validator class applies
+validator.WidgetValidatorDiscriminators(PodFormatsValidator, field=IConfigurablePODTemplate['pod_formats'])
+
+# Register the validator so it will be looked up by z3c.form machinery
+provideAdapter(PodFormatsValidator)
+
 
 class ConfigurablePODTemplate(PODTemplate):
     """
@@ -176,6 +216,7 @@ class ConfigurablePODTemplate(PODTemplate):
     """
 
     implements(IConfigurablePODTemplate)
+    meta_type = 'ConfigurablePODTemplate'
 
     def get_style_template(self):
         """
@@ -240,3 +281,11 @@ class SubTemplate(PODTemplate):
     """
 
     implements(ISubTemplate)
+    meta_type = 'SubTemplate'
+
+POD_TEMPLATE_TYPES = {
+    PODTemplate.__name__: PODTemplate,
+    ConfigurablePODTemplate.__name__: ConfigurablePODTemplate,
+    SubTemplate.__name__: SubTemplate,
+    StyleTemplate.__name__: StyleTemplate,
+}