+++ src/collective/documentgenerator/tests/test_AT_helper_view.py
@@ -173,10 +173,10 @@ class TestArchetypesHelperViewMethods(ArchetypesIntegrationTests):
 
         self._test_display(field_name, expected, result)
 
-    def test_display_text_method_without_appy_renderer(self):
+    def test_display_text_method(self):
         field_name = 'description'
-        to_set = 'Yolo!'
-        expected = ''
+        to_set = 'My description\r\nMy life\r\nhttp://www.imio.be'
+        expected = 'My description<br />My life<br /><a href="http://www.imio.be" rel="nofollow">http://www.imio.be</a>'
 
         field = self.AT_topic.getField(field_name)
         field.set(self.AT_topic, to_set)
@@ -184,9 +184,20 @@ class TestArchetypesHelperViewMethods(ArchetypesIntegrationTests):
         result = self.view.display_text(field_name)
         self._test_display(field_name, expected, result)
 
+    def test_display_html_method_without_appy_renderer(self):
+        field_name = 'description'
+        to_set = 'Yolo!'
+        expected = ''
+
+        field = self.AT_topic.getField(field_name)
+        field.set(self.AT_topic, to_set)
+
+        result = self.view.display_html(field_name)
+        self._test_display(field_name, expected, result)
+
     def test_check_permission(self):
         # test user has permission
-        self.assertTrue(self.view.check_permission('description', self.AT_topic))
+        self.assertTrue(self.view.check_permission('description'))
 
         # set field read permission to higher stuff
         field = self.AT_topic.getField('description')
@@ -194,4 +205,4 @@ class TestArchetypesHelperViewMethods(ArchetypesIntegrationTests):
         # new user that doesn't have permission
         api.user.create(username='foobar', email='foobar@example.com')
         login(self.portal, 'foobar')
-        self.assertFalse(self.view.check_permission('description', self.AT_topic))
+        self.assertFalse(self.view.check_permission('description'))