+++ src/collective/documentgenerator/tests/test_pod_template.py
@@ -4,10 +4,12 @@ from Acquisition import aq_base
 
 from collective.documentgenerator.content.condition import PODTemplateCondition
 from collective.documentgenerator.content.merge_templates import TemplatesToMergeForPODTemplate
+from collective.documentgenerator.content.pod_template import IConfigurablePODTemplate
 from collective.documentgenerator.content.pod_template import IPODTemplate
 from collective.documentgenerator.interfaces import IPODTemplateCondition
 from collective.documentgenerator.interfaces import ITemplatesToMerge
 from collective.documentgenerator.testing import PODTemplateIntegrationTest
+from collective.documentgenerator.testing import ConfigurablePODTemplateIntegrationTest
 from collective.documentgenerator.testing import TEST_INSTALL_INTEGRATION
 
 from plone import api
@@ -16,9 +18,14 @@ from plone.namedfile.file import NamedBlobFile
 from zope.component import getGlobalSiteManager
 from zope.component import queryAdapter
 from zope.component import queryMultiAdapter
+from zope.i18n import translate
 from zope.interface import Interface
+from zope.interface import Invalid
+
 
 import unittest
+
+from collective.documentgenerator.content.pod_template import PodFormatsValidator
 
 
 class TestPODTemplate(unittest.TestCase):
@@ -58,13 +65,13 @@ class TestPODTemplateFields(PODTemplateIntegrationTest):
         msg = "field 'odt_file' is not displayed"
         self.assertTrue('id="form-widgets-odt_file"' in contents, msg)
         msg = "field 'odt_file' is not translated"
-        self.assertTrue('Document odt' in contents, msg)
+        self.assertTrue('Canevas' in contents, msg)
 
     def test_odt_file_field_edit(self):
         self.browser.open(self.test_podtemplate.absolute_url() + '/edit')
         contents = self.browser.contents
         msg = "field 'odt_file' is not editable"
-        self.assertTrue('Document odt' in contents, msg)
+        self.assertTrue('Canevas' in contents, msg)
 
     def test_initial_md5_attribute(self):
         test_podtemplate = aq_base(self.test_podtemplate)
@@ -175,3 +182,30 @@ class TestPODTemplateIntegration(PODTemplateIntegrationTest):
 
     def test_get_available_formats(self):
         self.assertEqual(self.test_podtemplate.get_available_formats(), ['odt', ])
+
+
+class TestPODTemplateValidator(ConfigurablePODTemplateIntegrationTest):
+
+    def test_file_extension_is_odt(self):
+        pod_template = self.test_podtemplate
+        extension = pod_template.odt_file.filename.split('.')[-1]
+        self.assertEqual(extension, 'odt')
+
+    def test_add_bad_formats_and_get_errormessage(self):
+        pod_template = self.test_podtemplate
+        pod_template.pod_formats.append('xls')
+        view = pod_template.restrictedTraverse('edit')
+        view.form_instance.update()
+        validator = PodFormatsValidator(pod_template,
+                                        pod_template.REQUEST,
+                                        view.form_instance,
+                                        IConfigurablePODTemplate['pod_formats'],
+                                        pod_template.widget)
+        msg = translate(u"element_not_valid",
+                        default=u"Element ${elem} is not valid for .${extension} template : \"${template}\"",
+                        mapping={u"elem": "Microsoft Excel (.xls)",
+                                 u"extension": "odt",
+                                 u"template": pod_template.odt_file.filename})
+        with self.assertRaises(Invalid) as cm:
+            validator.validate(pod_template.pod_formats)
+        self.assertEqual(msg, translate(cm.exception.message))