+++ src/collective/documentgenerator/viewlets/generationlinks.py
@@ -31,17 +31,26 @@ class DocumentGeneratorLinksViewlet(ViewletBase):
 
         return generable_templates
 
+    def get_generation_view_name(self, template, output_format):
+        return 'document-generation'
+
     def get_links_info(self):
         base_url = self.context.absolute_url()
         links = []
         for template in self.get_generable_templates():
             for output_format in template.get_available_formats():
                 title = template.Title()
+                description = template.Description()
                 uid = template.UID()
-                link = '{base_url}/document-generation?template_uid={uid}&output_format={output_format}'.format(
+                link = '{base_url}/{gen_view_name}?template_uid={uid}&output_format={output_format}'.format(
                     base_url=base_url,
                     uid=uid,
                     output_format=output_format,
+                    gen_view_name=self.get_generation_view_name(template, output_format)
                 )
-                links.append({'link': link, 'title': title, 'output_format': output_format, 'template_uid': uid})
+                links.append({'link': link,
+                              'title': title,
+                              'description': description,
+                              'output_format': output_format,
+                              'template_uid': uid})
         return links